+++ examples/realworld/ext/auth/SignupPage.js
@@ -21,7 +21,7 @@ const SignupPage = () => {
     setSubmitError(null)
     try {
       await signup({ username, email, password })
-      await login(email, password)
+      await login(username, password)
       history.push('/')
     } catch (err) {
       setSubmitError(err)