+++ examples/thoughts/migrations/20220818165835_getting_up_to_date/migration.sql
@@ -1,0 +1,17 @@
+-- DropForeignKey
+ALTER TABLE "Tag" DROP CONSTRAINT "Tag_userId_fkey";
+
+-- DropForeignKey
+ALTER TABLE "Thought" DROP CONSTRAINT "Thought_userId_fkey";
+
+-- AddForeignKey
+ALTER TABLE "Thought" ADD CONSTRAINT "Thought_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "Tag" ADD CONSTRAINT "Tag_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+
+-- RenameIndex
+ALTER INDEX "Tag.name_userId_unique" RENAME TO "Tag_name_userId_key";
+
+-- RenameIndex
+ALTER INDEX "User.username_unique" RENAME TO "User_username_key";