+++ examples/tutorials/ItWaspsOnMyMachine/ext/actions.js
@@ -1,0 +1,5 @@
+export const saveExcuse = async (excuse, context) => {
+  return context.entities.Excuse.create({
+    data: { text: excuse.text }
+  })
+}