+++ examples/waspleau/ext/workers/utils.js
@@ -1,11 +1,9 @@
-import Prisma from '@prisma/client'
-
-const prisma = new Prisma.PrismaClient()
-
-export function upsertMetric({ name, value } = {}) {
-  return prisma.metric.upsert({
-    where: { name },
-    update: { name, value: String(value) },
-    create: { name, value: String(value) }
-  })
+export function upsertMetric(context) {
+  return ({ name, value } = {}) => {
+    return context.entities.Metric.upsert({
+      where: { name },
+      update: { name, value: String(value) },
+      create: { name, value: String(value) }
+    })
+  }
 }