+++ waspc/data/Generator/templates/server/src/jobs/_job.js
@@ -1,4 +1,5 @@
 {{={= =}=}}
+import prisma from '../dbClient.js'
 import { createJob } from './{= executorJobRelFP =}'
 {=& jobPerformFnImportStatement =}
 
@@ -6,5 +7,10 @@ export const {= jobName =} = createJob({
   jobName: "{= jobName =}",
   jobFn: {= jobPerformFnName =},
   defaultJobOptions: {=& jobPerformOptions =},
-  jobSchedule: {=& jobSchedule =}
+  jobSchedule: {=& jobSchedule =},
+  entities: {
+    {=# entities =}
+    {= name =}: prisma.{= prismaIdentifier =},
+    {=/ entities =}
+  },
 })