+++ waspc/examples/todoApp/ext/Todo.js
@@ -60,7 +60,6 @@ const Footer = (props) => {
       await deleteCompletedTasks()
     } catch (err) {
       console.log(err)
-      window.alert('Error:' + err.message)
     }
   }
 
@@ -99,9 +98,9 @@ const Tasks = (props) => {
 }
 
 const Task = (props) => {
-  const action = useAction(updateTaskIsDone, {
+  const updateTaskIsDoneOptimistically = useAction(updateTaskIsDone, {
     optimisticUpdates: [{
-      getQuery: () => [getTasks],
+      getQuerySpecifier: () => [getTasks],
       updateQuery: (updatedTask, oldTasks) => {
         if (oldTasks === undefined) {
           // cache is empty
@@ -117,10 +116,9 @@ const Task = (props) => {
     const isDone = event.target.checked
 
     try {
-      await action.mutateAsync({ id, isDone })
+      await updateTaskIsDoneOptimistically({ id, isDone })
     } catch (err) {
       console.log(err)
-      window.alert('Error:' + err.message)
     }
   }
 
@@ -157,7 +155,6 @@ const NewTaskForm = (props) => {
       setDescription(defaultDescription)
     } catch (err) {
       console.log(err)
-      window.alert('Error:' + err.message)
     }
   }
 
@@ -184,7 +181,6 @@ const ToggleAllTasksButton = (props) => {
       await toggleAllTasks()
     } catch (err) {
       console.log(err)
-      window.alert('Error:' + err.message)
     }
   }