+++ waspc/examples/todoApp/ext/actions.js
@@ -5,7 +5,7 @@ const sleep = (ms) => new Promise(res => setTimeout(res, ms))
 
 export const createTask = async (task, context) => {
   if (!context.user) {
-    throw new HttpError(403)
+    throw new HttpError(401)
   }
 
   const Task = context.entities.Task
@@ -24,7 +24,7 @@ export const createTask = async (task, context) => {
 
 export const updateTaskIsDone = async ({ id, isDone }, context) => {
   if (!context.user) {
-    throw new HttpError(403)
+    throw new HttpError(401)
   }
 
   await sleep(3000);
@@ -37,7 +37,7 @@ export const updateTaskIsDone = async ({ id, isDone }, context) => {
 
 export const deleteCompletedTasks = async (args, context) => {
   if (!context.user) {
-    throw new HttpError(403)
+    throw new HttpError(401)
   }
 
   const Task = context.entities.Task
@@ -48,7 +48,7 @@ export const deleteCompletedTasks = async (args, context) => {
 
 export const toggleAllTasks = async (args, context) => {
   if (!context.user) {
-    throw new HttpError(403)
+    throw new HttpError(401)
   }
 
   const whereIsDone = isDone => ({ isDone, user: { id: context.user.id } })