+++ waspc/src/Wasp/Generator/ServerGenerator/JobGenerator.hs
@@ -26,6 +26,7 @@ import StrongPath
   )
 import qualified StrongPath as SP
 import Wasp.AppSpec (AppSpec, getJobs)
+import qualified Wasp.AppSpec as AS
 import qualified Wasp.AppSpec.App.Dependency as AS.Dependency
 import qualified Wasp.AppSpec.JSON as AS.JSON
 import Wasp.AppSpec.Job (Job, JobExecutor (PgBoss, Simple), jobExecutors)
@@ -62,7 +63,8 @@ genJob (jobName, job) =
             -- `Aeson.Text.encodeToLazyText` on an Aeson.Object, or `show` on an AS.JSON.
             "jobSchedule" .= Aeson.Text.encodeToLazyText (fromMaybe Aeson.Null maybeJobSchedule),
             "jobPerformOptions" .= show (fromMaybe AS.JSON.emptyObject maybeJobPerformOptions),
-            "executorJobRelFP" .= toFilePath (executorJobTemplateInJobsDir (J.executor job))
+            "executorJobRelFP" .= toFilePath (executorJobTemplateInJobsDir (J.executor job)),
+            "entities" .= maybe [] (map (buildEntityData . AS.refName)) (J.entities job)
           ]
     )
   where
@@ -77,6 +79,13 @@ genJob (jobName, job) =
           "options" .= fromMaybe AS.JSON.emptyObject (J.scheduleExecutorOptionsJson job)
         ]
     maybeJobSchedule = jobScheduleTmplData <$> J.schedule job
+
+    buildEntityData :: String -> Aeson.Value
+    buildEntityData entityName =
+      object
+        [ "name" .= entityName,
+          "prismaIdentifier" .= C.entityNameToPrismaIdentifier entityName
+        ]
 
 -- Creates a file that is imported on the server to ensure all job JS modules are loaded
 -- even if they are not referenced by user code. This ensures schedules are started, etc.
@@ -137,7 +146,7 @@ jobsDirInServerRootDir = SP.castRel jobsDirInServerTemplatesDir
 -- NOTE: Our pg-boss related documentation references this version in URLs.
 -- Please update the docs when this changes (until we solve: https://github.com/wasp-lang/wasp/issues/596).
 pgBossVersionRange :: SV.Range
-pgBossVersionRange = SV.rangeFromVersion (SV.BackwardsCompatibleWith, SV.Version 7 2 1)
+pgBossVersionRange = SV.Range [SV.backwardsCompatibleWith (SV.Version 7 2 1)]
 
 pgBossDependency :: AS.Dependency.Dependency
 pgBossDependency = AS.Dependency.make ("pg-boss", show pgBossVersionRange)